configure: make sure the Pi firmware library path is searched first
Some libraries would include -L/usr/lib in their link flags. Because
link flags were appended to the beginning of LDFLAGS, this resulted in
/usr/lib being searched before /opt/vc/lib.

Normally this isn't a problem, because there are no EGL and GLESv2 in
/usr/lib only /opt/vc/lib. However, Arch includes system EGL and GLESv2
in /usr/lib. This results in them being linked against instead of the R-Pi's
firmware provided libraries.